pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel"
version = "0.1.0"
description = "AI-Native Security Platform with Zero-Trust Architecture and Autonomous Network Management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "1450 Enterprises LLC", email = "dev@1450enterprises.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: Security",
]
keywords = ["security", "networking", "zero-trust", "ai", "automation"]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "httpx>=0.25.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "redis>=5.0.1",
    "networkx>=3.2.1",
    "scapy>=2.5.0",
    "netifaces>=0.11.0",
    "schedule>=1.2.1",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
gui = [
    "PySide6>=6.6.0",
]

[project.scripts]
sentinel = "sentinel.cli.main:main"
sentinel-gui = "sentinel.gui.app:main"

[project.urls]
Homepage = "https://github.com/1450enterprises/sentinel"
Documentation = "https://sentinel.1450enterprises.com/docs"
Repository = "https://github.com/1450enterprises/sentinel"
Issues = "https://github.com/1450enterprises/sentinel/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sentinel = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=sentinel --cov-report=term-missing"

[tool.coverage.run]
source = ["src/sentinel"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
